Added coordinate tests for DST
git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@178236 16731396-06f5-0310-8873-f7f720988828

+++ dst/private/dst/HealPixCoordinates.cxx
@@ -102,7 +102,7 @@ uint32_t HealPixCoordinate::GetIndex(double theta, double phi)
 	return floatpair(theta,phi);
 }
 
-uint8_t HealPixCoordinate::NumberOfBins() 
+uint32_t HealPixCoordinate::NumberOfBins() 
 { 
 	HealpixMapDbl& skyMap = *( (HealpixMapDbl*) skyMap_);
     return skyMap.Npix();
